figured out quest implementation in comments

+++ app/src/test/java/com/example/calculatorprojectv2/QuestUnitTest.java
@@ -23,7 +23,7 @@ public class QuestUnitTest {
     public void quest_buttonLimit_isWorking() {
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            int[] testNums = q.getTestNums();
+            double[] testNums = q.getTestNums();
             // Checks to see if the length of each test number plus the length of the other test number plus one is the button limit
             // This is good for everything else
             assertEquals(String.valueOf(testNums[0]).length() + String.valueOf(testNums[1]).length() + 1,
@@ -35,7 +35,7 @@ public class QuestUnitTest {
     public void quest_targetNumber_isWorking() {
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            int[] testNums = q.getTestNums();
+            double[] testNums = q.getTestNums();
             switch (q.getOperationDesignation()) {
                 case 1:
                     assertEquals(testNums[0] + testNums[1], q.getTargetNumber());
@@ -65,7 +65,7 @@ public class QuestUnitTest {
     public void quest_numberLimit_isBelow() {
         for (int i = 0; i < 10000; i++) {
             Quest q = new Quest(i);
-            int[] testNums = q.getTestNums();
+            double[] testNums = q.getTestNums();
             assertTrue(testNums[0] < 10000);
             assertTrue(testNums[1] < 10000);
             assertTrue(q.getTargetNumber() < 10000);
